+++ examples/markdown-shortcuts/index.js
@@ -1,4 +1,4 @@
-import { Editor } from 'slate-react'
+import { Editor } from 'slate-react-41'
 import { Value } from 'slate'
 
 import React from 'react'