+++ packages/slate-react-41/test/index.js
@@ -5,7 +5,7 @@ import clean from './helpers/clean'
 import React from 'react'
 import ReactDOM from 'react-dom/server'
 import Simulator from 'slate-simulator'
-import { Editor } from 'slate-react'
+import { Editor } from 'slate-react-41'
 import { fixtures } from 'slate-dev-test-utils'
 import { JSDOM } from 'jsdom'